requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-bootstrap-form==3.4
django-datatable==0.3.1
django-djconfig==0.8.0
django-filter==2.1.0
django-haystack==2.8.1
django-import-export==1.2.0
django-infinite-scroll-pagination==0.2.0
django-js-asset==1.2.1
django-machina==0.7.1
django-mptt==0.9.1
django-spirit==0.7.0
django-tables2==2.0.4
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
jdcal==1.4
markdown2==2.3.7
mistune==0.7.4
mysqlclient==1.3.12
odfpy==1.4.0
olefile==0.45.1
openpyxl==2.5.14
Pillow==5.2.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
tablib==0.12.1
uni-slugify==0.1.4
unicodecsv==0.14.1
Unidecode==1.0.22
webstack-django-sorting==1.0.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
